feat(pyd): rename pyb module -> pyd
